+++ client/src/components/JobPost/ApplyConfirmModal.jsx
@@ -1,22 +1,46 @@
-import React from 'react';
-import { Button, Header, Image, Modal } from 'semantic-ui-react';
+import React, { Component } from 'react';
+import { Button, Confirm } from 'semantic-ui-react';
+import { Modal, OverlayTrigger } from 'react-bootstrap';
 
-const ApplyConfirmModal = () => {
-  return (
-    <div>
-      <Modal trigger={<Button>Apply!</Button>}>
-        <Modal.Header>Select a Photo</Modal.Header>
-        <Modal.Content image>
-          <Image wrapped size='medium' src='/assets/images/avatar/large/rachel.png' />
-          <Modal.Description>
-            <Header>Default Profile Image</Header>
-            <p>We've found the following gravatar image associated with your e-mail address.</p>
-            <p>Is it okay to use this photo?</p>
-          </Modal.Description>
-        </Modal.Content>
-      </Modal>
-    </div>
-  );
-};
 
-export default ApplyConfirmModal;
+class ApplyConfirm extends Component {
+  constructor(props) {
+    super(props);
+    this.state = {
+      showModal: false
+    };
+    this.show = this.show.bind(this);
+    this.handleConfirm = this.handleConfirm.bind(this);
+    this.handleClose = this.handleClose.bind(this);
+  }
+  show() {
+    this.setState({ showModal: true });
+  }
+  handleConfirm() {
+    this.setState({ showModal: false });
+  }
+  handleClose() {
+    this.setState({ showModal: false });
+  }
+  render() {
+    const { id, onJobInterviewClick } = this.props;
+    return (
+      <div>
+        <Button onClick={this.show} color="green" fluid>Apply!</Button>
+        <Modal show={this.state.showModal} onHide={this.handleClose}>
+          <Modal.Header closeButton>
+            <Modal.Title>Terms & Conditions</Modal.Title>
+          </Modal.Header>
+          <Modal.Body>
+            <h4>Honor Code: </h4>
+            <p>Duis mollis, est non commodo luctus, nisi erat porttitor ligula.</p>
+            <hr />
+            <Button onClick={() => { onJobInterviewClick(id); }} color="blue" attached="bottom">GO!</Button>
+          </Modal.Body>
+        </Modal>
+      </div>
+    );
+  }
+}
+
+export default ApplyConfirm;
