+++ package.json
@@ -25,6 +25,7 @@
     "auth0-lock": "^10.14.0",
     "axios": "^0.15.3",
     "body-parser": "^1.17.1",
+    "brace": "^0.10.0",
     "cors": "^2.8.2",
     "dotenv": "^4.0.0",
     "express": "^4.15.2",
@@ -32,6 +33,7 @@
     "jwt-decode": "^2.2.0",
     "lodash": "^4.17.4",
     "react": "^15.4.2",
+    "react-ace": "^4.1.6",
     "react-bootstrap": "^0.30.8",
     "react-bootstrap-select": "^0.14.0",
     "react-dom": "^15.4.2",
